Step R30: ComplexJson, the last family -- a re-parent that cashes in the hand copy (#56)
ComplexPropertiesQueryInfoCarrierFixture becomes ComplexJsonQueryInfoCarrierFixture on
ComplexJsonRelationalFixtureBase, and the seven classes move from ComplexProperties*TestBase to
ComplexJson*RelationalTestBase. Adopted bare: every override removed, so the failures are
measured before anything is written to answer them.

The ~20 lines of ToJson() C0 mirrored by hand are deleted, and the copy was diffed against the
original before it went: byte-identical apart from the wording of one comment.

This is a re-parent and not an addition, and running both would be duplication rather than
coverage. The ComplexJson*RelationalTestBase classes derive from the ComplexProperties*TestBase
ones, so compliance resolves both transitively, and the non-JSON complex mapping is not lost:
R28 adopted it as ComplexTableSplitting. Two complex mappings, one family each, and no model
mirrored by hand anywhere in the block any more.

Measured locally: Passed: 126, Failed: 10, Total: 136 -- and 136 is exactly the count the family
had before the re-parent, so no test is gained or lost. All ten are five ComplexJsonProjection
tests times two arms, all raising SqliteStrings.ApplyNotSupported bare in both arms, because a
complex type is not tracked as an entity and no owned-tracking assertion intervenes. R30a is the
override subset.

It also corrects a C0-era remark that stood on this file, which said the ComplexJson* bases
"assert SQL and stay unadopted". They do not, and the whole block is the evidence.

And it settles the #62 note the old file carried. That note deleted an override of
Contains_with_nested_and_composed_operators -- borrowed from
ComplexTableSplittingStructuralEqualityRelationalTestBase and applied to a JSON-mapped model --
and called the result a query this provider answers that other EF providers refuse. The sharper
reading now available: EF's JSON structural-equality base asserts nothing at all, so passing
there is agreement rather than divergence, and R28 runs the table-splitting base where EF does
assert the throw and this provider throws. The difference was the mapping, not the provider.
Deleting the override was right for a better reason than the one recorded.
